LayoutTests/fast/dom/Window/window-property-shadowing-onclick-expected.txt

This tests the ES5.2 behavior where global variables should not trigger setters on the Window object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL onClickCalled should be false. Was true.
